Initialise static HTML
Initialise static HTML in public folder with hand-crafted HTML and CSS.
The background image was created by midjourney. The reason to put the
contents in a public folder rather than the root of the repo is so that
we can add a README.md to the root that will not be part of the public
page. Additionally this setup allows for later static site generation.